backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
garminconnect>=0.2.36
anthropic==0.40.0
pydantic>=2.12.0
pydantic-settings>=2.1.0

